feat: improve edit album card styling

+++ src/components/Admin/EditPhotoDataCard.tsx
@@ -5,6 +5,8 @@ import {
     updatePhotoData,
 } from '../../firebase/firebase-firestore'
 import { PhotoData } from '../../types'
+import { TextArea, TextField, Label } from '../Form/Text'
+import { Button } from '../Buttons/Button'
 
 import './editPhotoDataCard.css'
 
@@ -26,47 +28,55 @@ const EditPhotoDataCard = ({
                 className="edit-photo-data-card__image"
                 src={photo.thumbnailUrl}
             />
-            <h2>{photo.fileName}</h2>
+            <h2 className="edit-photo-data-card__photo-name">
+                {photo.fileName}
+            </h2>
             <div className="edit-photo-data-card__input">
-                <label htmlFor="photoTitle">Title</label>
-                <input
-                    type="text"
+                <Label htmlFor="photoTitle">Title</Label>
+                <TextField
                     id="photoTitle"
                     value={photoTitle}
                     onChange={(e) => setPhotoTitle(e.target.value)}
                 />
             </div>
             <div className="edit-photo-data-card__input">
-                <label htmlFor="photoDescription">Description</label>
-                <textarea
+                <Label htmlFor="photoDescription">Description</Label>
+                <TextArea
                     id="photoDescription"
                     value={photoDescription}
                     onChange={(e) => setPhotoDescription(e.target.value)}
                 />
             </div>
-            <button
-                type="button"
-                onClick={() =>
-                    updatePhotoData(photo.documentRef, {
-                        title: photoTitle,
-                        description: photoDescription,
-                    })
-                }
-            >
-                Oppdater
-            </button>
-            <button
-                type="button"
-                onClick={async () => {
-                    await deleteImageFromStorage({
-                        albumName,
-                        fileName: photo.fileName,
-                    })
-                    window.location.reload()
-                }}
-            >
-                SLETT
-            </button>
+            <div className="edit-photo-data-card__buttons">
+                <Button
+                    onClick={() =>
+                        updatePhotoData(photo.documentRef, {
+                            title: photoTitle,
+                            description: photoDescription,
+                        })
+                    }
+                >
+                    Oppdater
+                </Button>
+                <Button
+                    variant="negative"
+                    onClick={async () => {
+                        if (
+                            window.confirm(
+                                `Sikker på at du vil slette ${photo.fileName}?`,
+                            )
+                        ) {
+                            await deleteImageFromStorage({
+                                albumName,
+                                fileName: photo.fileName,
+                            })
+                            window.location.reload()
+                        }
+                    }}
+                >
+                    Slett
+                </Button>
+            </div>
         </div>
     )
 }
